Moved some types to own modules
 - Expr lives in Dhall.Syntax.Expr
 - Import and friends live in Dhall.Syntax.Import
 - Const lives in Dhall.Syntax.Const
 - Var lives in Dhall.Syntax.Var
