+++ dhall/dhall.cabal
@@ -360,6 +360,9 @@ Library
         Dhall.Parser.Combinators
         Dhall.Pretty.Internal
         Dhall.Syntax
+        Dhall.Syntax.Const
+        Dhall.Syntax.Expr
+        Dhall.Syntax.Import
         Dhall.Syntax.Instances.Applicative
         Dhall.Syntax.Instances.Bifunctor
         Dhall.Syntax.Instances.Data
@@ -377,6 +380,7 @@ Library
         Dhall.Syntax.Instances.Traversable
         Dhall.Syntax.Operations
         Dhall.Syntax.Types
+        Dhall.Syntax.Var
         Dhall.URL
         Paths_dhall
     Autogen-Modules:
