+++ dhall/src/Dhall/Syntax.hs
@@ -2,6 +2,9 @@ module Dhall.Syntax
     ( module Export
     ) where
 
+import Dhall.Syntax.Const                 as Export
+import Dhall.Syntax.Expr                  as Export
+import Dhall.Syntax.Import                as Export
 import Dhall.Syntax.Instances.Applicative as Export ()
 import Dhall.Syntax.Instances.Bifunctor   as Export ()
 import Dhall.Syntax.Instances.Data        as Export ()
@@ -19,3 +22,4 @@ import Dhall.Syntax.Instances.Show        as Export ()
 import Dhall.Syntax.Instances.Traversable as Export ()
 import Dhall.Syntax.Operations            as Export
 import Dhall.Syntax.Types                 as Export
+import Dhall.Syntax.Var                   as Export
