--- dhall/src/Dhall/Syntax.hs-boot
@@ -1,7 +1,0 @@
-module Dhall.Syntax where
-
-data Var
-
-data Const
-
-data Expr s a
