+++ dhall/src/Dhall/Syntax/Expr.hs-boot
@@ -1,0 +1,3 @@
+module Dhall.Syntax.Expr where
+
+data Expr s a
