+++ dhall/src/Dhall/Syntax/Instances/Applicative.hs
@@ -2,6 +2,7 @@
 
 module Dhall.Syntax.Instances.Applicative () where
 
+import Dhall.Syntax.Expr
 import Dhall.Syntax.Operations
 import Dhall.Syntax.Types
