+++ dhall/src/Dhall/Syntax/Instances/Functor.hs
@@ -5,6 +5,7 @@
 
 module Dhall.Syntax.Instances.Functor () where
 
+import                Dhall.Syntax.Expr
 import {-# SOURCE #-} Dhall.Syntax.Operations (unsafeSubExpressions)
 import                Dhall.Syntax.Types
