+++ dhall/src/Dhall/Syntax/Instances/NFData.hs
@@ -2,8 +2,12 @@
 
 module Dhall.Syntax.Instances.NFData () where
 
-import Control.DeepSeq    (NFData)
+import Control.DeepSeq     (NFData)
+import Dhall.Syntax.Const
+import Dhall.Syntax.Expr
+import Dhall.Syntax.Import
 import Dhall.Syntax.Types
+import Dhall.Syntax.Var
 
 instance NFData Const
 instance NFData Var
