+++ dhall/src/Dhall/Syntax/Instances/Ord.hs
@@ -4,8 +4,12 @@
 
 module Dhall.Syntax.Instances.Ord () where
 
+import Dhall.Syntax.Const
+import Dhall.Syntax.Expr
+import Dhall.Syntax.Import
 import Dhall.Syntax.Instances.Eq ()
 import Dhall.Syntax.Types
+import Dhall.Syntax.Var
 
 deriving instance Ord Const
 deriving instance Ord Var
