+++ dhall/src/Dhall/Syntax/Instances/Traversable.hs
@@ -5,6 +5,7 @@
 
 module Dhall.Syntax.Instances.Traversable () where
 
+import Dhall.Syntax.Expr
 import Dhall.Syntax.Instances.Foldable ()
 import Dhall.Syntax.Instances.Functor  ()
 import Dhall.Syntax.Types
