+++ dhall/src/Dhall/Syntax/Operations.hs-boot
@@ -1,5 +1,5 @@
 module Dhall.Syntax.Operations where
 
-import Dhall.Syntax.Types (Expr)
+import Dhall.Syntax.Expr (Expr)
 
 unsafeSubExpressions :: Applicative f => (Expr s a -> f (Expr t b)) -> Expr s a -> f (Expr t b)
