+++ dhall/src/Dhall/Syntax/Types.hs-boot
@@ -1,7 +1,17 @@
 module Dhall.Syntax.Types where
 
-data Var
+data Binding s a
 
-data Const
+data DhallDouble
 
-data Expr s a
+data Chunks s a
+
+data PreferAnnotation s a
+
+data RecordField s a
+
+data FunctionBinding s a
+
+data FieldSelection s
+
+data WithComponent
